refactor: flatten config type hierarchy, remove dead abstractions
Config system cleanup based on full dependency analysis:

- Remove CoreContext (dead — never called by any production code)
- Remove CoreGraphQLOptions (over-engineered, 11 fields defined but only 4 populated)
- Flatten CoreConfig: security/federation/codegen/externalServices as direct fields
  instead of nested in a never-fully-populated CoreGraphQLOptions bag
- Remove FrameworkAdapter.createCoreContext (dead interface method)
- Make ExternalGraphQLService extend ExternalServiceCodegenConfig (was fully
  redeclaring all fields without type relationship)
- Make FederationConfig extend CoreFederationConfig (was parallel hierarchy)
- Deduplicate LocalDirExtendSource: defined in core, re-exported from nitro
- Remove mergeGraphQLOptions (unused utility)
- Fix CLI buildDir default: `.nitro-graphql` → `.graphql` (matches DEFAULT_CLI_CONFIG)
- Simplify both adapters to pass config fields directly

Before: 3 parallel type hierarchies (ExternalService, Federation, Paths)
After: single inheritance chain from core → nitro
