+++ src/core/config.ts
@@ -1,14 +1,13 @@
 /**
  * Core configuration utilities
- * Factory functions and helpers for creating CoreConfig
+ * Factory functions for creating CoreConfig and ScanContext
  */
 
-import type { CoreConfig, CoreContext, CoreGraphQLOptions, CoreLogger } from './types/config'
+import type { CoreConfig, CoreLogger } from './types/config'
 import type { ScanContext } from './types/scanning'
-import { relative, resolve } from 'pathe'
+import { resolve } from 'pathe'
 import {
   DIR_APP_GRAPHQL,
-  DIR_BUILD_NITRO,
   DIR_CLIENT_GRAPHQL,
   DIR_SERVER_GRAPHQL,
   GRAPHQL_FRAMEWORK_YOGA,
@@ -19,24 +18,18 @@ import { createLogger } from './utils/logger'
  * Options for creating a CoreConfig
  */
 export interface CreateCoreConfigOptions {
-  /** Root directory of the project */
   rootDir: string
-  /** Build directory (optional, defaults based on framework) */
   buildDir?: string
-  /** Server GraphQL directory (optional) */
   serverDir?: string
-  /** Client GraphQL directory (optional) */
   clientDir?: string
-  /** Whether running in Nuxt context */
   isNuxt?: boolean
-  /** Whether running in development mode */
   isDev?: boolean
-  /** GraphQL options */
-  graphqlOptions?: CoreGraphQLOptions
-  /** Custom logger */
   logger?: CoreLogger
-  /** Patterns to ignore */
   ignorePatterns?: string[]
+  security?: CoreConfig['security']
+  federation?: CoreConfig['federation']
+  codegen?: CoreConfig['codegen']
+  externalServices?: CoreConfig['externalServices']
 }
 
 /**
@@ -47,16 +40,15 @@ export function createCoreConfig(options: CreateCoreConfigOptions): CoreConfig {
     rootDir,
     isNuxt = false,
     isDev = process.env.NODE_ENV !== 'production',
-    graphqlOptions = {},
     logger = createLogger(),
     ignorePatterns = [],
+    security,
+    federation,
+    codegen,
+    externalServices,
   } = options
 
-  // Determine framework
-  const framework = graphqlOptions.framework || GRAPHQL_FRAMEWORK_YOGA
-
-  // Resolve directories
-  const buildDir = options.buildDir || resolve(rootDir, isNuxt ? '.nuxt' : DIR_BUILD_NITRO)
+  const buildDir = options.buildDir || resolve(rootDir, isNuxt ? '.nuxt' : '.nitro')
   const serverDir = options.serverDir || resolve(rootDir, DIR_SERVER_GRAPHQL)
   const clientDir = options.clientDir || resolve(rootDir, isNuxt ? DIR_APP_GRAPHQL : DIR_CLIENT_GRAPHQL)
   const typesDir = resolve(buildDir, 'types')
@@ -67,33 +59,15 @@ export function createCoreConfig(options: CreateCoreConfigOptions): CoreConfig {
     serverDir,
     clientDir,
     typesDir,
-    framework,
+    framework: GRAPHQL_FRAMEWORK_YOGA,
     isNuxt,
     isDev,
-    graphqlOptions,
     logger,
     ignorePatterns,
-  }
-}
-
-/**
- * Create a CoreContext from CoreConfig
- */
-export function createCoreContext(config: CoreConfig): CoreContext {
-  const graphqlBuildDir = resolve(config.buildDir, 'graphql')
-
-  // Calculate relative directory paths
-  const dir = {
-    build: relative(config.rootDir, config.buildDir),
-    client: relative(config.rootDir, config.clientDir),
-    server: relative(config.rootDir, config.serverDir),
-  }
-
-  return {
-    config,
-    graphqlBuildDir,
-    watchDirs: [],
-    dir,
+    security,
+    federation,
+    codegen,
+    externalServices,
   }
 }
 
@@ -108,30 +82,5 @@ export function createScanContext(config: CoreConfig): ScanContext {
     ignorePatterns: config.ignorePatterns,
     isDev: config.isDev,
     logger: config.logger,
-  }
-}
-
-/**
- * Merge GraphQL options with defaults
- */
-export function mergeGraphQLOptions(
-  options: Partial<CoreGraphQLOptions>,
-  defaults: Partial<CoreGraphQLOptions> = {},
-): CoreGraphQLOptions {
-  return {
-    ...defaults,
-    ...options,
-    codegen: {
-      ...defaults.codegen,
-      ...options.codegen,
-    },
-    security: {
-      ...defaults.security,
-      ...options.security,
-    },
-    paths: {
-      ...defaults.paths,
-      ...options.paths,
-    },
   }
 }
